requirements.txt
asgiref==3.8.1attrs==24.2.0autobahn==24.4.2Automat==24.8.1cachetools==5.5.0certifi==2024.8.30cffi==1.17.1channels==4.1.0chardet==5.2.0charset-normalizer==3.3.2cloudinary==1.41.0constantly==23.10.4cryptography==43.0.1daphne==4.1.2Django==5.1.1django-allauth==65.1.0django-cloudinary-storage==0.3.0django-cors-headers==4.4.0django-mailer==2.3.2djangorestframework==3.15.2djangorestframework-simplejwt==5.3.1google-auth==2.35.0hyperlink==21.0.0idna==3.10incremental==24.7.2lockfile==0.12.2pillow==10.4.0psycopg2==2.9.9pyasn1==0.6.1pyasn1_modules==0.4.1pycparser==2.22PyJWT==2.9.0pyOpenSSL==24.2.1PyPDF2==3.0.1python-dotenv==1.0.1reportlab==4.2.2requests==2.32.3rsa==4.9service-identity==24.1.0six==1.16.0sqlparse==0.5.1Twisted==24.7.0txaio==23.1.1typing_extensions==4.12.2tzdata==2024.1urllib3==2.2.3whitenoise==6.7.0zope.interface==7.0.3